# packages/contracts/openapi/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Media Analysis Platform API",
    "version": "2.0.0",
    "description": "Single-user channel-aware target contract for API-owned media assets, collections, immutable selection snapshots, analysis runs, artifacts, diagnostics, channel accounts, channel surfaces, and idempotency."
  },
  "paths": {
...
    "/v1/media-assets/{media_asset_id}/exports": {
      "post": {
        "operationId": "createExportJob",
        "summary": "Create an idempotent one-material export job using a semantic variant.",
        "parameters": [
          {"$ref": "#/components/parameters/MediaAssetId"},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "../schemas/http/export-job.schema.json#/$defs/createExportJobRequest"}}}
        },
        "responses": {
          "201": {
            "description": "Export job created or replayed.",
            "content": {"application/json": {"schema": {"$ref": "../schemas/http/export-job.schema.json#/$defs/exportJobResponse"}}}
          },
          "409": {"$ref": "#/components/responses/ErrorEnvelope"}
        }
      }
    },
...
    "/v1/export-jobs": {
      "get": {
        "operationId": "listExportJobs",
        "summary": "List export jobs scoped to one channel account.",
        "parameters": [
          {"$ref": "#/components/parameters/ChannelAccountId"},
          {"$ref": "#/components/parameters/PageCursor"},
          {"$ref": "#/components/parameters/PageSize"}
        ],
        "responses": {
          "200": {
            "description": "Export job page.",
            "content": {"application/json": {"schema": {"$ref": "../schemas/http/export-job.schema.json#/$defs/exportJobListResponse"}}}
          }
        }
      }
    },
    "/v1/export-jobs/{export_job_id}": {
      "get": {
        "operationId": "getExportJob",
        "summary": "Read one channel-account-scoped export job.",
        "parameters": [
          {"$ref": "#/components/parameters/ChannelAccountId"},
          {"$ref": "#/components/parameters/ExportJobId"}
        ],
        "responses": {
          "200": {
            "description": "Export job state.",
            "content": {"application/json": {"schema": {"$ref": "../schemas/http/export-job.schema.json#/$defs/exportJobResponse"}}}
          },
          "404": {"$ref": "#/components/responses/ErrorEnvelope"}
        }
      }
    },
    "/v1/export-jobs/{export_job_id}/cancel": {
      "post": {
        "operationId": "cancelExportJob",
        "summary": "Cancel queued work immediately or request cooperative active cancellation.",
        "parameters": [{"$ref": "#/components/parameters/ExportJobId"}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "../schemas/http/export-job.schema.json#/$defs/cancelExportJobRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Updated export job.",
            "content": {"application/json": {"schema": {"$ref": "../schemas/http/export-job.schema.json#/$defs/exportJobResponse"}}}
          },
          "409": {"$ref": "#/components/responses/ErrorEnvelope"}
        }
      }
    },
    "/v1/export-jobs/{export_job_id}/retry": {
      "post": {
        "operationId": "retryExportJob",
        "summary": "Explicitly retry a failed or canceled export with a new generation.",
        "parameters": [
          {"$ref": "#/components/parameters/ExportJobId"},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "../schemas/http/export-job.schema.json#/$defs/retryExportJobRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Requeued export job.",
            "content": {"application/json": {"schema": {"$ref": "../schemas/http/export-job.schema.json#/$defs/exportJobResponse"}}}
          },
          "409": {"$ref": "#/components/responses/ErrorEnvelope"}
        }
      }
    },
...
    "/v1/export-jobs/{export_job_id}/download": {
      "get": {
        "operationId": "resolveExportDownload",
        "summary": "Resolve short-lived account-scoped output access without acknowledging delivery.",
        "parameters": [
          {"$ref": "#/components/parameters/ChannelAccountId"},
          {"$ref": "#/components/parameters/ExportJobId"}
        ],
        "responses": {
          "200": {
            "description": "Short-lived attachment access.",
            "content": {"application/json": {"schema": {"$ref": "../schemas/http/export-job.schema.json#/$defs/exportDownloadResponse"}}}
          },
          "404": {"$ref": "#/components/responses/ErrorEnvelope"}
        }
      }
    },
...
    "/internal/v1/export-jobs/{export_job_id}/claim": {
      "post": {
        "operationId": "claimExportJob",
        "summary": "Acquire a new export attempt and lease fence.",
        "parameters": [{"$ref": "#/components/parameters/ExportJobId"}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "../schemas/internal/export-control.schema.json#/$defs/exportClaimRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Claimed export job and source access.",
            "content": {"application/json": {"schema": {"$ref": "../schemas/internal/export-control.schema.json#/$defs/exportClaimResponse"}}}
          },
          "409": {"$ref": "#/components/responses/ErrorEnvelope"}
        }
      }
    },
...
    "/internal/v1/export-jobs/{export_job_id}/finalize": {
      "post": {
        "operationId": "finalizeExportJob",
        "summary": "Fenced export finalization with output lineage or safe diagnostics.",
        "parameters": [{"$ref": "#/components/parameters/ExportJobId"}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "../schemas/internal/export-control.schema.json#/$defs/exportFinalizeRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Terminal export state.",
            "content": {"application/json": {"schema": {"$ref": "../schemas/http/export-job.schema.json#/$defs/exportJobResponse"}}}
          },
          "409": {"$ref": "#/components/responses/ErrorEnvelope"}
        }
      }
    },
    "/internal/v1/export-jobs/{export_job_id}/download-access": {
      "get": {
        "operationId": "resolveInternalExportDownloadAccess",
        "summary": "Resolve private service-to-service access to a succeeded export output.",
        "parameters": [
          {"$ref": "#/components/parameters/ChannelAccountId"},
          {"$ref": "#/components/parameters/ExportJobId"}
        ],
        "responses": {
          "200": {
            "description": "Short-lived private attachment access for an authenticated adapter.",
            "content": {"application/json": {"schema": {"$ref": "../schemas/http/export-job.schema.json#/$defs/exportDownloadResponse"}}}
          },
          "404": {"$ref": "#/components/responses/ErrorEnvelope"}
        }
      }
    },
    "/internal/v1/export-jobs/reclaim": {
      "post": {
        "operationId": "reclaimExportJobs",
        "summary": "Requeue expired export leases or fail exhausted jobs.",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "../schemas/internal/export-control.schema.json#/$defs/exportReclaimRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Reclaim summary.",
            "content": {"application/json": {"schema": {"$ref": "../schemas/internal/export-control.schema.json#/$defs/exportReclaimResponse"}}}
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "schema": {
          "$ref": "../schemas/common/enums.schema.json#/$defs/idempotencyKey"
        }
      },
...
      "PageCursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "../schemas/common/pagination.schema.json#/$defs/pageCursor"
        }
      },
      "PageSize": {
        "name": "page_size",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "../schemas/common/pagination.schema.json#/$defs/pageSize"
        }
      },
      "ChannelAccountId": {
        "name": "channel_account_id",
        "in": "query",
        "required": true,
        "schema": {
          "$ref": "../schemas/common/enums.schema.json#/$defs/uuid"
        }
      },
...
      "MediaAssetId": {
        "name": "media_asset_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../schemas/common/enums.schema.json#/$defs/uuid"
        }
      },
...
      "ExportJobId": {
        "name": "export_job_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../schemas/common/enums.schema.json#/$defs/uuid"
        }
      },
...
    },
    "responses": {
      "ErrorEnvelope": {
        "description": "Validation, channel-scope, conflict, retention, or diagnostic error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/error-envelope.schema.json"
            }
          }
        }
      }
    },
...
  }
}